CompressingReader: compressed data reader stream (#210)
In Go, `http.Request` requires an `io.ReadCloser` for an upload body. To
facilitate compression of data being uploaded we want to wrap a
source reader (such an `io.File`) with a proposed
`lz4.CompressingReader`, whereupon the said reader can be set directly
into `http.Request` Body field.

This benefits not only the plain `http.Request`, but also the various
network file storage APIs which build on top of Go's http client.

Signed-off-by: Alex Dubov <oakad@yahoo.com>